Make sure that valid dates aren't NaT'd in Doganella aquifer clean
Because dates can't be centered/scaled, we have to save a reference
to the old date series and add it back onto the data post-scaling.
But in the process of scaling, we create a brand new dataframe--
with a brand new index.

If we do not remove the old index from the date series, then the
conflicting sets of indexes can cause some of the dates to be
garbled into NaT.

Prevent this.